Updated to allow use of production licence with launchInstall
Updated to allow use of production licence with launchInstall. Note that
it is now hardcoded to use vagrant-provision-auth file as default
production licence file but can easily be changed in the
docker-compose.yml file. Still not fully running end to end though as
the REST API doesnt seem to be set up since its looping in the rest api
request check.